requirements-knowledge.txt
# Knowledge Agent Dependencies (Prompt 04)
# Add these to your project's requirements.txt

# Core ML/AI
langgraph>=0.2.0
langchain-core>=0.3.0
sentence-transformers>=2.2.0
torch>=2.0.0  # Required by sentence-transformers

# Database
asyncpg>=0.29.0
pgvector>=0.2.0

# Web scraping
httpx>=0.25.0
beautifulsoup4>=4.12.0

# File processing
pypdf>=3.17.0
python-docx>=1.1.0

# Existing dependencies (if not already in your requirements.txt)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
cryptography>=41.0.0
PyJWT>=2.8.0
requests>=2.31.0
